Api-Backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
pandas
python-multipart
matplotlib